+++ src/releases/components/helpers/index.js
@@ -1,0 +1,1 @@
+export {default as ReleasesHelp} from './ReleasesHelp';
